# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Viveris Carbone API",
    "description": "FastAPI backend for Viveris Carbone project",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": [
          "users",
          "users"
        ],
        "summary": "Get user dashboard data",
        "description": "Returns user profile, total emissions, category breakdown, points, and achievements.",
        "operationId": "get_me_api_users_me_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDashboardResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users",
          "users"
        ],
        "summary": "Delete user account",
        "operationId": "delete_user_api_users_me_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hard",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Hard"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me/password": {
      "put": {
        "tags": [
          "users",
          "users"
        ],
        "summary": "Change user password",
        "operationId": "change_password_api_users_me_password_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/users/me/stats": {
      "get": {
        "tags": [
          "users",
          "users"
        ],
        "summary": "Get advanced user stats",
        "operationId": "get_advanced_stats_api_users_me_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdvancedUserStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/users/me/stats/recalculate": {
      "post": {
        "tags": [
          "users",
          "users"
        ],
        "summary": "Force recalculate stats",
        "operationId": "recalculate_stats_api_users_me_stats_recalculate_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericActionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/users/me/achievements/{achievementId}/unlock": {
      "post": {
        "tags": [
          "users",
          "users"
        ],
        "summary": "Manually unlock an achievement",
        "operationId": "unlock_achievement_api_users_me_achievements__achievementId__unlock_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "achievementId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Achievementid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/challenges/recommendations": {
      "get": {
        "tags": [
          "gamification"
        ],
        "summary": "Get recommended challenges",
        "description": "Returns a list of challenges recommended for the user, including completion status.",
        "operationId": "get_recommendations_api_challenges_recommendations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChallengeResponse"
                  },
                  "type": "array",
                  "title": "Response Get Recommendations Api Challenges Recommendations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/challenges/{challenge_id}/toggle": {
      "post": {
        "tags": [
          "gamification"
        ],
        "summary": "Toggle a challenge completion",
        "description": "Marks a challenge as completed or uncompleted. Awards or removes points and updates tree planting progress.",
        "operationId": "toggle_challenge_api_challenges__challenge_id__toggle_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "challenge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Challenge Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleChallengeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToggleChallengeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/challenges": {
      "post": {
        "tags": [
          "gamification"
        ],
        "summary": "Create a custom challenge",
        "description": "Creates a new custom challenge. Points must be between 5 and 500.",
        "operationId": "create_challenge_api_challenges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateChallengeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Achievement": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "unlocked": {
            "type": "boolean",
            "title": "Unlocked"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "unlocked"
        ],
        "title": "Achievement"
      },
      "AdvancedUserStatsResponse": {
        "properties": {
          "co2ThisYear": {
            "type": "number",
            "title": "Co2Thisyear"
          },
          "co2LastYear": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Co2Lastyear",
            "default": 0
          },
          "co2ReducedThisYear": {
            "type": "number",
            "title": "Co2Reducedthisyear"
          },
          "challengesCompleted": {
            "type": "integer",
            "title": "Challengescompleted"
          },
          "bestStreak": {
            "type": "integer",
            "title": "Beststreak"
          },
          "currentStreak": {
            "type": "integer",
            "title": "Currentstreak"
          },
          "badges": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Badges"
          }
        },
        "type": "object",
        "required": [
          "co2ThisYear",
          "co2ReducedThisYear",
          "challengesCompleted",
          "bestStreak",
          "currentStreak",
          "badges"
        ],
        "title": "AdvancedUserStatsResponse"
      },
...
      "ChallengeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "points",
          "category",
          "completed"
        ],
        "title": "ChallengeResponse"
      },
      "CreateChallengeRequest": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "title",
          "category",
          "points"
        ],
        "title": "CreateChallengeRequest"
      },
      "CreateChallengeResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "challenge": {
            "additionalProperties": true,
            "type": "object",
            "title": "Challenge"
          }
        },
        "type": "object",
        "required": [
          "challenge"
        ],
        "title": "CreateChallengeResponse"
      },
...
      "GenericActionResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "GenericActionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordChangeRequest": {
        "properties": {
          "currentPassword": {
            "type": "string",
            "title": "Currentpassword"
          },
          "newPassword": {
            "type": "string",
            "title": "Newpassword"
          }
        },
        "type": "object",
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "title": "PasswordChangeRequest"
      },
...
      "QuizResult": {
        "properties": {
          "totalInTons": {
            "type": "number",
            "title": "Totalintons"
          },
          "categoryBreakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Categorybreakdown"
          }
        },
        "type": "object",
        "required": [
          "totalInTons",
          "categoryBreakdown"
        ],
        "title": "QuizResult"
      },
...
      "ToggleChallengeRequest": {
        "properties": {
          "completed": {
            "type": "boolean",
            "title": "Completed"
          }
        },
        "type": "object",
        "required": [
          "completed"
        ],
        "title": "ToggleChallengeRequest"
      },
      "ToggleChallengeResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "challengeId": {
            "type": "string",
            "title": "Challengeid"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed"
          },
          "pointsDelta": {
            "type": "integer",
            "title": "Pointsdelta"
          },
          "totalPoints": {
            "type": "integer",
            "title": "Totalpoints"
          },
          "treesPlanted": {
            "type": "integer",
            "title": "Treesplanted"
          },
          "treeProgress": {
            "type": "integer",
            "title": "Treeprogress"
          }
        },
        "type": "object",
        "required": [
          "challengeId",
          "completed",
          "pointsDelta",
          "totalPoints",
          "treesPlanted",
          "treeProgress"
        ],
        "title": "ToggleChallengeResponse"
      },
      "UserObject": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "firstName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstname"
          },
          "lastName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastname"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "firstName",
          "lastName",
          "email"
        ],
        "title": "UserObject"
      },
...
      "UserProfileDashboardResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserObject"
          },
          "quizResult": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QuizResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "categoryEmissions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoryemissions"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "treesPlanted": {
            "type": "integer",
            "title": "Treesplanted"
          },
          "achievements": {
            "items": {
              "$ref": "#/components/schemas/Achievement"
            },
            "type": "array",
            "title": "Achievements"
          },
          "streak": {
            "type": "integer",
            "title": "Streak"
          },
          "bestStreak": {
            "type": "integer",
            "title": "Beststreak"
          }
        },
        "type": "object",
        "required": [
          "user",
          "points",
          "treesPlanted",
          "achievements",
          "streak",
          "bestStreak"
        ],
        "title": "UserProfileDashboardResponse",
        "example": {
          "achievements": [
            {
              "id": "first-tree",
              "name": "Premier arbre",
              "unlocked": true
            }
          ],
          "bestStreak": 18,
          "categoryEmissions": {
            "Alimentation": 1.2,
            "Consommation": 0.7,
            "Transport": 1.8,
            "Énergie": 1.5
          },
          "points": 1240,
          "quizResult": {
            "categoryBreakdown": {
              "Alimentation": 1.2,
              "Consommation": 0.7,
              "Transport": 1.8,
              "Énergie": 1.5
            },
            "totalInTons": 2.8
          },
          "streak": 12,
          "treesPlanted": 2,
          "user": {
            "email": "jean.dupont@example.com",
            "firstName": "Jean",
            "id": "uuid",
            "lastName": "Dupont"
          }
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}